Refactored Image-Upload
