+++ apps/quick-learn-frontend/src/app/(Dashboard)/dashboard/profile-settings/ProfileSettings.tsx
@@ -15,6 +15,8 @@ import React, { useContext, useEffect, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { z } from 'zod';
 import ProfileSettingsSkeleton from './ProfileSettingsSkeleton';
+import { FileUploadResponse } from '@src/shared/types/utilTypes';
+import { AxiosSuccessResponse } from '@src/apiServices/axios';
 
 // Define the service input type
 interface ProfileUpdateServiceInput {
@@ -98,7 +100,29 @@ const ProfileSettings = () => {
       disabled: true,
     },
   ];
-
+  //change umage only
+  function onChangeImage(
+    res: AxiosSuccessResponse<FileUploadResponse> | undefined,
+  ) {
+    const serviceInput = {
+      first_name: user?.first_name,
+      last_name: user?.last_name,
+      profile_image: res ? res.data.file : '', //handle upload and delete
+    };
+    if (user) {
+      setUser({
+        ...user,
+        profile_image: res ? res.data.file : '',
+      });
+    }
+    updateUserProfileService(serviceInput)
+      .then((response) => {
+        showApiMessageInToast(response);
+      })
+      .catch((err) => {
+        showApiErrorInToast(err);
+      });
+  }
   const onSubmit = async (data: ProfileFormData) => {
     setIsLoading(true);
     try {
@@ -162,6 +186,7 @@ const ProfileSettings = () => {
           onSubmit={onSubmit}
           methods={methods}
           buttonText="Save"
+          onChangeImage={onChangeImage}
           id="profileSettingsForm"
         />
       </FormProvider>
