+++ apps/quick-learn-frontend/src/shared/types/utilTypes.ts
@@ -7,3 +7,7 @@ export type PaginateWrapper<T> = {
   limit: number;
   total_pages: number;
 };
+export type FileUploadResponse = {
+  file: string;
+  type: string;
+};
